--renamed some parameters --changed easy logging

+++ include/frame.hpp
@@ -81,7 +81,7 @@ public:
     explicit Frame( const std::shared_ptr< PinholeCamera >& camera,
                     const cv::Mat& img,
                     const uint32_t maxImagePyramid,
-                    const double timestamp );
+                    const uint64_t timestamp );
     // Copy C'tor
     Frame( const Frame& rhs ) = delete;  // non construction-copyable
     // move C'tor
@@ -191,7 +191,7 @@ public:
     ImagePyramid m_imagePyramid;                      ///< Image pyramid of image
     std::vector< std::shared_ptr< Feature > > m_features;  ///< List of features in the image
     bool m_keyFrame;                                       ///< Is this frames selected as keyframe?
-    double m_timestamp;                                    ///< Timestamp of when the image was recorded
+    uint64_t m_timestamp;                                    ///< Timestamp of when the image was recorded
 
 private:
 };
